pyproject.toml
[tool.poetry]
name = "binge-platform"
version = "1.0.0"
description = "booking platform"
authors = ["Data engineering"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.95.2"
structlog = "^23.1.0"
gunicorn = "^20.1.0"
orjson = "^3.8.12"
uvicorn = "^0.22.0"
pybrake = "^1.10.1"
gremlinpython = "^3.6.3"
psycopg2-binary = "^2.9.6"
requests = "^2.31.0"
yoyo-migrations = "^8.2.0"
python-dotenv = "^1.0.1"
sib-api-v3-sdk = "^7.6.0"
fastapi-utilities = "^0.2.0"
jinja2 = "^3.1.4"
pdfkit = "^1.0.0"
boto3 = "^1.35.20"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.1"
black = "^23.3.0"
seed-isort-config = "^2.2.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
src_paths = ["isort", "test"]